Cleaning up config files. Condensed leaderhip and task into a single config. Moved the HttpSessionConfig into common location
