+++ genie-web/src/main/java/com/netflix/genie/web/configs/HttpSessionConfig.java
@@ -15,7 +15,7 @@
  *     limitations under the License.
  *
  */
-package com.netflix.genie.web.sessions;
+package com.netflix.genie.web.configs;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
